+++ module/system.js
@@ -187,6 +187,7 @@ Hooks.once("init", function () {
     "systems/palladium-universal/templates/item/parts/description.hbs",
     "systems/palladium-universal/templates/vehicle/time-machine.hbs",
     "systems/palladium-universal/templates/actor/parts/item-roll.hbs",
+    "systems/palladium-universal/templates/actor/parts/counter.hbs",
     "systems/palladium-universal/templates/actor/parts/air-combat.hbs"
   ]);
 });